requirements.txt
# -- Core AI & Quant Libraries --
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.2
statsmodels>=0.14.0

# -- Financial Data & Technical Analysis --
yfinance>=0.2.28
ta>=0.11.0  # Technical Analysis Library
matplotlib>=3.7.1
seaborn>=0.12.2
mplfinance>=0.12.9b7

# -- Bot & Async IO --
python-telegram-bot[job-queue]>=20.0
aiohttp>=3.8.0

# -- Web Dashboard Support (Optional/Later) --
fastapi>=0.95.0
uvicorn>=0.22.0
websockets>=11.0
